+++ mec-api/mec_voice.h
@@ -116,7 +116,7 @@ public:
             voice->vel_.raw_ = voice->vel_.scale_ *
                                (voice->vel_.x_ * voice->vel_.sumxy_ - (voice->vel_.sumx_ * voice->vel_.sumy_))
                                / (voice->vel_.x_ * voice->vel_.sumxsq_ - (voice->vel_.sumx_ * voice->vel_.sumx_));
-            voice->v_ = 1.0f - pow((double) (1.0f - voice->vel_.raw_), (double) (voice->vel_.curve_));
+            voice->v_ = static_cast<float>(1.0f - pow((double) (1.0f - voice->vel_.raw_), (double) (voice->vel_.curve_)));
 
             // LOG_1("vel detector : " << voice->id_);
             // LOG_1("raw : " << voice->vel_.raw_ << " v " << voice->v_);
